refactor: 공백 수정 및 상수화 및 테스트 코드 분리
