+++ src/main/java/nextstep/courses/domain/ImageType.java
@@ -1,9 +1,14 @@
 package nextstep.courses.domain;
 
+import java.util.EnumSet;
+import java.util.Set;
+
 public enum ImageType {
   GIF, JPG, JPEG, PNG, SVG;
 
+  private static final Set<ImageType> SUPPORTED_TYPES = EnumSet.of(GIF, JPG, JPEG, PNG, SVG);
+
   public boolean isSupported() {
-    return this == GIF || this == JPG || this == JPEG || this == PNG || this == SVG;
+    return SUPPORTED_TYPES.contains(this);
   }
 }
